+++ src/Support/helpers.php
@@ -247,7 +247,7 @@ function theme_favicon(): string
 
 function system_version(): string
 {
-    return '0.37.0';
+    return '0.38.0';
 }
 
 function system_label(): string